RedisClient updatePlace uses TTL does not expire
